Base classes restored and modified.
