+++ app/src/main/java/ar/com/wolox/android/bootstrap/ui/posts/PostsViewModel.kt
@@ -2,12 +2,11 @@ package ar.com.wolox.android.bootstrap.ui.posts
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import ar.com.wolox.android.bootstrap.Constants
 import ar.com.wolox.android.bootstrap.model.Post
-import ar.com.wolox.android.bootstrap.network.util.RequestStatus
 import ar.com.wolox.android.bootstrap.repository.PostRepository
+import ar.com.wolox.android.bootstrap.ui.base.BaseViewModel
 import ar.com.wolox.android.bootstrap.utils.SharedPreferencesManager
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -17,25 +16,7 @@ import javax.inject.Inject
 class PostsViewModel @Inject constructor(
     private val postsRepository: PostRepository,
     private val sharedPreferencesManager: SharedPreferencesManager
-) : ViewModel() {
-
-    private val _requestStatus = MutableLiveData<RequestStatus>()
-    val requestStatus: LiveData<RequestStatus>
-        get() = _requestStatus
-
-    private fun toggleRequestStatus() {
-        _requestStatus.apply {
-            value = if (value != RequestStatus.Loading) {
-                RequestStatus.Loading
-            } else {
-                RequestStatus.Finished
-            }
-        }
-    }
-
-    private fun onRequestFailed(error: Int) {
-        _requestStatus.value = RequestStatus.Failure(error)
-    }
+) : BaseViewModel() {
 
     private val _posts = MutableLiveData<List<Post>>()
     val posts: LiveData<List<Post>>
